nuke: correctly push main.dart to main.dart.sentry

+++ nuke.sh
@@ -48,14 +48,15 @@ if [[ "$YN" =~ ^[yY] ]]; then
   git tag -l | grep '+' | xargs git tag -d
 fi
 
-read -n 1 -r -p "Handle main.dart? (y/N) " YN
-echo
-if [[ "$YN" =~ ^[yY] ]]; then
-  cp -f lib/main.dart lib/main.dart.sentry
-  cp -f lib/main.dart.vanilla lib/main.dart
-  git add lib/main.dart*
-  git commit -m 'Handle main.dart'
-  git push
+if [[ -n "$SENTRY_DSN" ]]; then
+  read -n 1 -r -p "Push main.dart back to main.dart.sentry? (y/N) " YN
+  echo
+  if [[ "$YN" =~ ^[yY] ]]; then
+    cp -f lib/main.dart.sentry lib/main.dart.swap
+    cp -f lib/main.dart lib/main.dart.sentry
+    mv -f lib/main.dart.swap lib/main.dart
+    git add lib/main.dart*
+    git commit -m 'Handle main.dart'
+    git push
+  fi
 fi
-
-
